refactor eventcardshimmer into its own files
