+++ apps/web/src/components/EventCards/__tests__/EventCard.test.js
@@ -1,7 +1,7 @@
 import { ThemeProvider, createTheme } from '@mui/material/styles';
 import { render, screen } from '@testing-library/react';
 
-import EventCard, { EventCardSkeleton } from '../EventCard';
+import EventCard from '../EventCard';
 
 describe('EventCard', () => {
     it.each([['light'], ['dark']])('renders in %s mode', (themeMode) => {
@@ -21,24 +21,5 @@ describe('EventCard', () => {
 
         expect(screen.getByRole('article', { name: 'Event card' })).toBeInTheDocument();
         expect(screen.getByText('Test content')).toBeInTheDocument();
-    });
-});
-
-describe('EventCardSkeleton', () => {
-    it.each([['light'], ['dark']])('renders skeleton in %s mode', (themeMode) => {
-        const muiTheme = createTheme({
-            palette: {
-                mode: themeMode
-            }
-        });
-
-        render(
-            <ThemeProvider theme={muiTheme}>
-                <EventCardSkeleton />
-            </ThemeProvider>
-        );
-
-        expect(screen.getByTestId('event-card-shimmer')).toBeInTheDocument();
-        expect(screen.getByLabelText('Loading event card')).toBeInTheDocument();
     });
 });
